Backend completo

+++ backend/Controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require ('../models/tareasModel')     //instalar npm i express-asy
 
 const getTareas  = asyncHandler(async (req,res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id})
     res.status(200).json({tareas})
 
 })
@@ -16,7 +16,8 @@ const createTareas = asyncHandler(async (req,res) => {
     }
 
     const tarea = await Tarea.create({ 
-        descripcion: req.body.descripcion
+        descripcion: req.body.descripcion,
+        user: req.user.id
      })
 
     res.status(201).json({tarea})
@@ -31,9 +32,18 @@ const updateTareas = asyncHandler(async (req,res) => {
         throw new Error('Esa Tarea no existe')
     }
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+    // nos aseguramos que la tarea pertenezca al usuario logeado, es decir el del token
 
-    res.status(200).json({tareaUpdated})
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+
+        res.status(200).json({tareaUpdated}) 
+    }
+
+    
 })
 
 const deleteTareas = asyncHandler(async (req,res) => {
@@ -45,10 +55,19 @@ const deleteTareas = asyncHandler(async (req,res) => {
         throw new Error('Esa Tarea no existe')
     }
 
-    await Tarea.deleteOne(tarea)
- //   const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)    
+    // nos aseguramos que la tarea pertenezca al usuario logeado, es decir el del token
 
-    res.status(200).json({ id: req.params.id  })
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no autorizado')
+    } else {
+        await Tarea.deleteOne(tarea)
+        //   const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)    
+       
+        res.status(200).json({ id: req.params.id  })
+       
+    }
+
 })
